# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  dto.CreatePriceUnitRequest:
    properties:
      base_currency:
        type: string
      code:
        type: string
      conversion_rate:
        type: number
      name:
        type: string
      precision:
        maximum: 8
        minimum: 0
        type: integer
      symbol:
        maxLength: 10
        type: string
    required:
    - base_currency
    - code
    - conversion_rate
    - name
    - symbol
    type: object
...
  dto.ListPriceUnitsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.PriceUnitResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.PriceUnitResponse:
    properties:
      base_currency:
        type: string
      code:
        type: string
      conversion_rate:
        type: number
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      precision:
        type: integer
      status:
        $ref: '#/definitions/types.Status'
      symbol:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdatePriceUnitRequest:
    properties:
      conversion_rate:
        type: number
      name:
        type: string
      precision:
        maximum: 8
        minimum: 0
        type: integer
      symbol:
        maxLength: 10
        type: string
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
  gin.H:
    additionalProperties: {}
    type: object
...
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.Status:
    enum:
    - published
    - deleted
    - archived
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDeleted
    - StatusArchived
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
...
  /prices/units:
    get:
      consumes:
      - application/json
      description: Get a paginated list of price units with optional filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListPriceUnitsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List price units
      tags:
      - Price Units
    post:
      consumes:
      - application/json
      description: Create a new price unit with the provided details
      parameters:
      - description: Price unit details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePriceUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PriceUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new price unit
      tags:
      - Price Units
  /prices/units/{id}:
    delete:
      consumes:
      - application/json
      description: Archive an existing price unit. The unit will be marked as archived
        and cannot be used in new prices.
      parameters:
      - description: Price unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Archive a price unit
      tags:
      - Price Units
    get:
      consumes:
      - application/json
      description: Get a price unit by ID
      parameters:
      - description: Price unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PriceUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a price unit by ID
      tags:
      - Price Units
    put:
      consumes:
      - application/json
      description: Update an existing price unit with the provided details. Only name,
        symbol, precision, and conversion_rate can be updated. Status changes are
        not allowed.
      parameters:
      - description: Price unit ID
        in: path
        name: id
        required: true
        type: string
      - description: Price unit details to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePriceUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PriceUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a price unit
      tags:
      - Price Units
  /prices/units/code/{code}:
    get:
      consumes:
      - application/json
      description: Get a price unit by code
      parameters:
      - description: Price unit code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PriceUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a price unit by code
      tags:
      - Price Units
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"